CHERRY-PICK Add ExecutableStage -> ProcessBundleDescriptor
This is the default conversion of some executable stage to the
ProcessBunProcessBundleDescriptor expected by the Control Service
client.

Add CoderTranslation#knownCoderOrByteArrayCoder

This converts a coder into a known coder (one with a registered
translator) or into a byte[] coder, recursively applied to any component
coders as required.

This is used by runners to instantiate wire coders without having to
mutate any components.

Close Outstanding Control Clients on Service close

FnApiControlClients can't outlive the service that provided them, so
when that service closes it should signal all existing SDK harnesses
that it is shutting down.

Clean up GrpcFnServer Constructor Signature

It can't throw an IOException, so stop declaring that it does. Factory
methods still may.

Add a RemoteExecutionTest

This takes an ExecutableStage and from a Java Pipeline and executes it
over the Beam Portability framework.

Clean up a couple of SdkHarnessClient Signatures

Make things not-final, remove checked exceptions that aren't thrown from
that method.