Add SdkHarnessManager singleton implementation (#5)
The Flink Runner uses a static SdkHarnessManager that remains for the
duration of a JVM process.  The SdkHarnessManager is responsible for
creating and managing resources from the java-fn-execution on behalf of
short-lived flink operators.

Usually this would be split across multiple smaller commits.  However,
we chose to squash these changes to simplify an anticipated cherry-pick
operation.  Changes introduced by this commit are:

* Add an implementation of the ArtifactRetrievalService called
  ArtifactProxyService which uses an ArtifactSource

* Make RemoteEnvironment public for runners

* SdkHarnessManager Singleton Implementation