Add ExecutableStagePayload to make aid runner stage reconstruction
Runner code generally only has access to the root pipeline and
`AppliedPTransform`s while walking the pipeline to perform translation.
This change adds a new `ExecutableStagePayload` message which is
embedded in the `FunctionSpec` of serialized `ExecutableStage`
`PTransform`s. Doing so avoids the extra pipeline munging necessary to
access to proper set of subtransforms.

It also means that subtransforms are _not_ added directly to
`ExecutableStage` transforms. This effectively makes executable stages
native transforms, which they should be as far as the runner is
concerned.